Update permit script to use arbitrary number of RWTs in commitments
Previously, watchers used just one RWT to create a commitment, but now
they may need to use more than 1 RWT. The number of required RWT is
announced in the RWTConfig box for each network. Watchers need to check
this number offline and use the required number of RWTs. Unless, their
commitments are useless in trigger creation and they have to redeem the
commitment.
In commitment creation we have to keep at least 1 RWT token in permit.
In other words, the permit box is not an optional output box in this
transaction.