Add RWT config to RWT repo Boxes
We wanted to create a new RWT config box to contain all related configs.
Although this splitting made the config adjustment easier, we needed the
number of watchers in trigger creation. Thus, the splitting is going to
be pointless when we need both boxes as data input. Then, all required
configs added to the RwtRepo.