Fix RWT check on permit box creating new events
Previously, while creating reports using permits, the RWT can be mixed
with arbitrary input token and be stored in other token places rather
than the first place. Now we check the RWT to be stored in the first
token place of both permit and commitment. Additionally, we check that
all input RWT tokens exist in permit and commitment box output boxes.