Fix commitment tx fee box evaluation error
To check the RWT tokens to exist in outputs in right place with the
correct amount, we have calculated the total input permits. To
distinguish the permit boxes we filtered the boxes with RWT at the first
place. But if we use fee boxes in this transaction that doesn't have any
token in it, results in the evaluation error for that box. Then we added
an extra condition to filter the boxes containing token.