Add repo config NFT
To minimize the transactions requiring the spenditure of guard NFT, the
repo config contract updated to have the guard NFT as data input and
receive required signs for update. It also has a new NFT now that makes
it unique from all other boxes. This nft is unique for each chain config.