Refactor scripts to store permit count in collateral box
Previously, rwt repo kept track of all watcher wids and permit count
itself, but in this update the watcher permit count is migrated to
collateral box. Now the collateral box is responsible for updating the
total permit count of the watcher, and it's added to other transactions
to update this value.
All current unit tests updated to match the current changes.